Restructure the files to clean BFT module
